camel-master/requirements-dev.txt
-r requirements-base.txt
ipython==2.3.1
django-debug-toolbar==1.3.0
